# server/src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: Magic Internal API
  version: "1.0"

paths:
  /server/start:
    post:
      summary: "POST server/start"
      operationId: "startGame"
      parameters:
        - name: playerOneName
          in: query
          required: true
          description: player one name
          schema:
            type: string
        - name: playerTwoName
          in: query
          required: true
          description: player two name
          schema:
            type: string
      responses:
        200:
          $ref: "#/components/responses/ServerResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /server/game/player:
    get:
      summary: "GET server/game/player"
      operationId: "getPlayerHealthAndMana"
      parameters:
        - name: playerName
          in: query
          required: true
          description: player name
          schema:
            type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerBean"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /server/game/player/active:
    get:
      summary: "GET server/game/player/active"
      operationId: "getActivePlayer"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerActiveBean"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
    put:
      summary: "PUT server/game/player/active"
      operationId: "setActivePlayer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerActiveBean"
      responses:
        200:
          $ref: "#/components/responses/ServerResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /server/game/player/play:
    put:
      summary: "POST server/game/player/play"
      operationId: "playCard"
      parameters:
        - name: playerName
          in: query
          required: true
          description: player name
          schema:
            type: string
        - name: cardIndex
          in: query
          required: true
          description: card index to play
          schema:
            type: number
      responses:
        200:
          $ref: "#/components/responses/ServerResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /server/game/player/card:
    get:
      summary: "GET server/game/player/card"
      operationId: "getPlayerCards"
      parameters:
        - name: playerName
          in: query
          required: true
          description: player name
          schema:
            type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DamageCardBean"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
components:
  responses:
    ServerResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseBean"
    BadRequest:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBean"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBean"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBean"
  schemas:
    PlayerBean:
      type: object
      properties:
        name:
          type: string
          description: "player name"
        health:
          type: integer
          description: "player health"
        manaSlots:
          type: integer
          description: "player mana slots"
      required:
        - name
    PlayerActiveBean:
      type: object
      properties:
        name:
          type: string
          description: "active player name"
      required:
        - name
    DamageCardBean:
      type: object
      properties:
        mana:
          type: string
          description: "card mana cost"
      required:
        - mana
    ResponseBean:
      type: object
      properties:
        message:
          type: string
          description: "Response message"
      required:
        - message
    ErrorBean:
      type: object
      properties:
        errorCode:
          type: integer
          description: "Error code, same value as the HTTP status, or dedicated values for some specific errors."
          enum:
            - 400
            - 404
            - 500
            - 1401
          x-enum-varnames:
            - BAD_PARAMETER
            - NOT_FOUND
            - INTERNAL_ERROR
            - GAME_OVER
        message:
          type: string
          description: "Error message"
      required:
        - errorCode
        - message